web-app/cars-be/requirements.txt
absl-py==0.11.0
aiofiles==0.5.0
aniso8601==7.0.0
astunparse==1.6.3
async-exit-stack==1.0.1
async-generator==1.10
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==2.0.0
email-validator==1.1.2
fastapi==0.63.0
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
grpcio==1.32.0
h11==0.12.0
h5py==2.10.0
httptools==0.1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Preprocessing==1.1.2
Markdown==3.3.3
MarkupSafe==1.1.1
numpy==1.19.4
oauthlib==3.1.0
opt-einsum==3.3.0
orjson==3.4.6
Pillow==8.1.0
promise==2.3
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.7.3
python-dotenv==0.15.0
python-multipart==0.0.5
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
Rx==1.6.1
six==1.15.0
starlette==0.13.6
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
typing-extensions==3.7.4.3
ujson==3.2.0
urllib3==1.26.2
uvicorn==0.13.3
uvloop==0.14.0
watchgod==0.6
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
